chore: update Dockerfile with stage without VOLUME definition
The VOLUME definition may not be needed in most cases, and
could clutter up diskspace since volumes are note deleted
by default when container exits. It was decided to leave
the volume in the default target of the Dockerfile, but
create a previous stage without VOLUME that is used for
building image for dockerhub.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>